* business:
  Merge code bug fixes
  Withdraw Giftcard functionality
  Beneficial Owners Changes
  Beneficial Owners Changes
  Transaction Details
  Beneficial Owners Changes
  ID verification Add bank
  Scan Activity ,DBA changes and ReceivePayment
  Team Validations
  Merchant related Bugfixes in sign in flow.
  Team Validations
  Beneficial Owners Upload doc and patch call
  Profile DBAInfo , Receive Payments ,Transaction List
  Beneficial Owners Upload doc and patch call
  Team Validations
  Bug Fixes
  Profile DBAInfo , Receive Payments ,Transaction List
  Profile DBAInfo , Receive Payments ,Transaction List
  BusinessAccount - Recent Transaction List

# Conflicts:
#	app/src/main/java/com/greenbox/coyni/view/BindingLayoutActivity.java
#	app/src/main/java/com/greenbox/coyni/view/business/BusinessRegistrationTrackerActivity.java

+++ app/src/main/java/com/greenbox/coyni/model/businesswallet/WalletName.java
@@ -8,7 +8,7 @@ public class WalletName {
     private int availabilityToUse;
     private int exchangeRate;
     private String exchangeVariation;
-    private int exchangeAmount;
+    private double exchangeAmount;
     private String status;
     private String walletCategory;
     private String type;
@@ -74,11 +74,11 @@ public class WalletName {
         this.exchangeVariation = exchangeVariation;
     }
 
-    public int getExchangeAmount() {
+    public double getExchangeAmount() {
         return exchangeAmount;
     }
 
-    public void setExchangeAmount(int exchangeAmount) {
+    public void setExchangeAmount(double exchangeAmount) {
         this.exchangeAmount = exchangeAmount;
     }
